test(llm): move call tests to module mirrors
