+++ tests/integrations/langchain/test_langchain_llm_adapter.py
@@ -18,9 +18,11 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 from langchain_core.messages import AIMessage, AIMessageChunk, HumanMessage, SystemMessage, ToolMessage
 
+from nemoguardrails.exceptions import LLMCallException
 from nemoguardrails.integrations.langchain.llm_adapter import (
     LangChainFramework,
     LangChainLLMAdapter,
+    _infer_provider_from_module,
     _langchain_chunk_to_llm_response_chunk,
     _langchain_response_to_llm_response,
 )
@@ -28,7 +30,164 @@ from nemoguardrails.integrations.langchain.message_utils import (
     chatmessage_to_langchain_message,
     chatmessages_to_langchain_messages,
 )
+from nemoguardrails.llm.call import llm_call
 from nemoguardrails.types import ChatMessage, LLMModel, LLMResponse, LLMResponseChunk, Role, ToolCall, ToolCallFunction
+
+
+class MockOpenAILLM:
+    __module__ = "langchain_openai.chat_models"
+
+
+class MockAnthropicLLM:
+    __module__ = "langchain_anthropic.chat_models"
+
+
+class MockNVIDIALLM:
+    __module__ = "langchain_nvidia_ai_endpoints.chat_models"
+
+
+class MockCommunityOllama:
+    __module__ = "langchain_community.chat_models.ollama"
+
+
+class MockUnknownLLM:
+    __module__ = "some_custom_package.models"
+
+
+def test_infer_provider_openai():
+    llm = MockOpenAILLM()
+    provider = _infer_provider_from_module(llm)
+    assert provider == "openai"
+
+
+def test_infer_provider_anthropic():
+    llm = MockAnthropicLLM()
+    provider = _infer_provider_from_module(llm)
+    assert provider == "anthropic"
+
+
+def test_infer_provider_nvidia_ai_endpoints():
+    llm = MockNVIDIALLM()
+    provider = _infer_provider_from_module(llm)
+    assert provider == "nvidia_ai_endpoints"
+
+
+def test_infer_provider_community_ollama():
+    llm = MockCommunityOllama()
+    provider = _infer_provider_from_module(llm)
+    assert provider == "ollama"
+
+
+def test_infer_provider_unknown():
+    llm = MockUnknownLLM()
+    provider = _infer_provider_from_module(llm)
+    assert provider is None
+
+
+def test_infer_provider_checks_base_classes():
+    class BaseOpenAI:
+        __module__ = "langchain_openai.chat_models"
+
+    class CustomWrapper(BaseOpenAI):
+        __module__ = "my_custom_wrapper.llms"
+
+    llm = CustomWrapper()
+    provider = _infer_provider_from_module(llm)
+    assert provider == "openai"
+
+
+def test_infer_provider_multiple_inheritance():
+    class BaseNVIDIA:
+        __module__ = "langchain_nvidia_ai_endpoints.chat_models"
+
+    class Mixin:
+        __module__ = "some_mixin.utils"
+
+    class MultipleInheritance(Mixin, BaseNVIDIA):
+        __module__ = "custom_package.models"
+
+    llm = MultipleInheritance()
+    provider = _infer_provider_from_module(llm)
+    assert provider == "nvidia_ai_endpoints"
+
+
+def test_infer_provider_deeply_nested_inheritance():
+    class Original:
+        __module__ = "langchain_anthropic.chat_models"
+
+    class Wrapper1(Original):
+        __module__ = "wrapper1.models"
+
+    class Wrapper2(Wrapper1):
+        __module__ = "wrapper2.models"
+
+    class Wrapper3(Wrapper2):
+        __module__ = "wrapper3.models"
+
+    llm = Wrapper3()
+    provider = _infer_provider_from_module(llm)
+    assert provider == "anthropic"
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize("llm_params", [None, {}])
+async def test_llm_call_stop_tokens_passed_without_llm_params(llm_params):
+    mock_llm = MagicMock()
+    mock_llm.model_name = "gpt-4"
+    bound_llm = AsyncMock()
+    bound_llm.ainvoke.return_value = MagicMock(content="response")
+    mock_llm.bind.return_value = bound_llm
+
+    wrapped = LangChainLLMAdapter(mock_llm)
+    await llm_call(wrapped, "prompt", stop=["User:"], llm_params=llm_params)
+
+    mock_llm.bind.assert_called_once_with(stop=["User:"])
+
+
+@pytest.mark.asyncio
+async def test_llm_call_exception_enrichment_with_model_and_provider():
+    mock_llm = MockOpenAILLM()
+    mock_llm.model_name = "gpt-4"
+    mock_llm.ainvoke = AsyncMock(side_effect=ConnectionError("Connection refused"))
+
+    wrapped = LangChainLLMAdapter(mock_llm)
+    with pytest.raises(LLMCallException) as exc_info:
+        await llm_call(wrapped, "test prompt")
+
+    exc_str = str(exc_info.value)
+    assert "gpt-4" in exc_str
+    assert "provider=openai" in exc_str
+    assert "Connection refused" in exc_str
+    assert isinstance(exc_info.value.inner_exception, ConnectionError)
+
+
+@pytest.mark.asyncio
+async def test_llm_call_exception_without_provider():
+    mock_llm = MockUnknownLLM()
+    mock_llm.model_name = "custom-model"
+    mock_llm.ainvoke = AsyncMock(side_effect=ValueError("Invalid request"))
+
+    wrapped = LangChainLLMAdapter(mock_llm)
+    with pytest.raises(LLMCallException) as exc_info:
+        await llm_call(wrapped, "test prompt")
+
+    exc_str = str(exc_info.value)
+    assert "custom-model" in exc_str
+    assert "Invalid request" in exc_str
+
+
+@pytest.mark.asyncio
+async def test_llm_call_kwargs_flow_through_to_generate():
+    mock_llm = MagicMock()
+    mock_llm.model_name = "gpt-4"
+    bound_llm = AsyncMock()
+    bound_llm.ainvoke.return_value = MagicMock(content="response")
+    mock_llm.bind.return_value = bound_llm
+
+    wrapped = LangChainLLMAdapter(mock_llm)
+    await llm_call(wrapped, "prompt", llm_params={"temperature": 0.5, "max_tokens": 100})
+
+    mock_llm.bind.assert_called_once_with(temperature=0.5, max_tokens=100)
 
 
 class TestLangChainLLMAdapter:
